refactor: inline body serialization as private methods on Impit
